+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v6.0.0 (TBD)
 - Major UI redesign based on [Slint framework](https://slint.dev).
 - Dropped GTK4 dependency for generic builds. It is still required for `mobile-access` feature to show an embedded webview.
+- Refactored split-tunnel routing setup to allow the routes which contain the VPN gateway address.
 - Fixed a regression with the additional search domains used as routing domains.
 
 ## v5.3.0 (2026-04-11)